modularized code editor
