+++ src/styles/styles.css
@@ -69,5 +69,5 @@ body {
 }
 
 .CodeMirror {
-  height: 85vh !important;
+  height: 80vh !important;
 }
